Refactor: split oversized files, fix duplicate logic, and standardize naming #129
