+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@joshuafolkken/kit",
-	"version": "0.56.0",
+	"version": "0.57.0",
 	"type": "module",
 	"license": "MIT",
 	"packageManager": "pnpm@10.33.0+sha512.10568bb4a6afb58c9eb3630da90cc9516417abebd3fabbe6739f0ae795728da1491e9db5a544c76ad8eb7570f5c4bb3d6c637b2cb41bfdcdb47fa823c8649319",
@@ -81,8 +81,8 @@
 		"prettier-plugin-svelte": "^3.5.1",
 		"prettier-plugin-tailwindcss": "^0.7.2",
 		"typescript": "^6.0.3",
-		"typescript-eslint": "^8.58.2",
-		"vitest": "^4.1.4"
+		"typescript-eslint": "^8.59.0",
+		"vitest": "^4.1.5"
 	},
 	"dependencies": {
 		"strip-json-comments": "^5.0.3",
