+++ scripts/init.test.ts
@@ -45,19 +45,22 @@ describe('copy_sonar_file_write', () => {
 	})
 })
 
-const INIT_SOURCE = readFileSync(fileURLToPath(new URL('init.ts', import.meta.url)), 'utf8')
+const AI_COPY_SOURCE = readFileSync(
+	fileURLToPath(new URL('init-ai-copy.ts', import.meta.url)),
+	'utf8',
+)
 
 describe('skip messages', () => {
 	it('reference josh sync not pnpm sync', () => {
-		expect(INIT_SOURCE).not.toContain('pnpm sync')
+		expect(AI_COPY_SOURCE).not.toContain('pnpm sync')
 	})
 
 	it('contain josh sync in file skip message', () => {
-		expect(INIT_SOURCE).toContain('run josh sync to update')
+		expect(AI_COPY_SOURCE).toContain('run josh sync to update')
 	})
 
 	it('contain josh sync in summary tip message', () => {
-		expect(INIT_SOURCE).toContain('Run `josh sync`')
+		expect(AI_COPY_SOURCE).toContain('Run `josh sync`')
 	})
 })
