+++ templates/playwright/base.ts
@@ -1,21 +1,27 @@
 import { defineConfig, devices } from '@playwright/test'
 
-// 環境判定と設定値の定数化
-const isCI = Boolean(process.env['CI'])
-// const isStaging = Boolean(process.env['STAGING']) // 将来の拡張用
+const is_ci = Boolean(process.env['CI'])
 
 const DEV_PORT = 5173
 const PREVIEW_PORT = 4173
 const CI_TIMEOUT = 15_000
 const LOCAL_TIMEOUT = 25_000
 const TEST_TIMEOUT = 10_000
-const EXPECT_TIMEOUT = 5000
-const ACTION_TIMEOUT = 5000
+const EXPECT_TIMEOUT = 5_000
+const ACTION_TIMEOUT = 5_000
 const NAVIGATION_TIMEOUT = 10_000
+const CI_WORKERS = 2
+const CI_RETRIES = 2
+const VIEWPORT_WIDTH = 1_280
+const VIEWPORT_HEIGHT = 720
 
-// 環境に応じた設定を関数化
-const getWebServerConfig = () => {
-	if (isCI) {
+function get_web_server_config(): {
+	command: string
+	port: number
+	timeout: number
+	reuseExistingServer: boolean
+} {
+	if (is_ci) {
 		return {
 			command: 'pnpm run preview',
 			port: PREVIEW_PORT,
@@ -24,8 +30,6 @@ const getWebServerConfig = () => {
 		}
 	}
 
-	// 将来的に staging 環境を追加する場合
-	// if (isStaging) { ... }
 	return {
 		command: 'pnpm run dev',
 		port: DEV_PORT,
@@ -35,51 +39,33 @@ const getWebServerConfig = () => {
 }
 
 export default defineConfig({
-	webServer: getWebServerConfig(),
+	webServer: get_web_server_config(),
 	testDir: 'e2e',
 	fullyParallel: true,
-	// 並列実行数を最適化（CPU コア数に応じて自動調整、CI では明示的に設定）
-	...(isCI ? { workers: 2 } : {}),
-	// リトライ設定（CI でのみ有効、ローカルでは即座に失敗を確認）
-	retries: isCI ? 2 : 0,
-	// タイムアウト設定を最適化
+	...(is_ci ? { workers: CI_WORKERS } : {}),
+	retries: is_ci ? CI_RETRIES : 0,
 	timeout: TEST_TIMEOUT,
 	expect: {
 		timeout: EXPECT_TIMEOUT,
 	},
-	// 必要なブラウザのみ実行
 	projects: [
 		{
 			name: 'chromium',
 			use: {
 				...devices['Desktop Chrome'],
-				// パフォーマンス向上のための設定
-				viewport: { width: 1280, height: 720 },
-				// ブラウザの起動を高速化
+				viewport: { width: VIEWPORT_WIDTH, height: VIEWPORT_HEIGHT },
 				launchOptions: {
-					args: [
-						'--disable-dev-shm-usage',
-						'--disable-gpu',
-						// --no-sandbox は CI 環境でのみ使用（セキュリティ上の理由）
-						...(isCI ? ['--no-sandbox'] : []),
-					],
+					args: ['--disable-dev-shm-usage', '--disable-gpu', ...(is_ci ? ['--no-sandbox'] : [])],
 				},
 			},
 		},
 	],
-	// レポート設定
-	reporter: isCI ? [['html'], ['github']] : [['html'], ['list']],
-	// グローバル設定
+	reporter: is_ci ? [['html'], ['github']] : [['html'], ['list']],
 	use: {
-		// アクションのタイムアウト
 		actionTimeout: ACTION_TIMEOUT,
-		// ナビゲーションのタイムアウト
 		navigationTimeout: NAVIGATION_TIMEOUT,
-		// スクリーンショットは失敗時のみ（CI でのみ）
-		screenshot: isCI ? 'only-on-failure' : 'off',
-		// ビデオは失敗時のみ（CI でのみ）
-		video: isCI ? 'retain-on-failure' : 'off',
-		// トレースは失敗時のみ（CI でのみ）
-		trace: isCI ? 'retain-on-failure' : 'off',
+		screenshot: is_ci ? 'only-on-failure' : 'off',
+		video: is_ci ? 'retain-on-failure' : 'off',
+		trace: is_ci ? 'retain-on-failure' : 'off',
 	},
 })
